Minor improvements

+++ src/support/mytime.h
@@ -28,6 +28,6 @@ void add_time( struct timespec t1, struct timespec t2, struct timespec* sum );
 void mult_time( struct timespec t1, int factor, struct timespec* prod );
 void reset( struct timespec *t1 );
 double to_secs( struct timespec t1 );
-long long timeval_diff( const struct timeval *start, const struct timeval *finish );
+long timeval_diff( const struct timeval *start, const struct timeval *finish );
 
 #endif /* MYTIME_H */
